Fix image deletion and paths

+++ image_upload/database/crud.py
@@ -11,8 +11,10 @@ def create_image(db: Session, file_name: str, file_hash: str, user_id: int):
     db.refresh(db_image)
     return db_image
 
+
 def delete_image(db: Session, image_id: int):
     db_image = db.query(models.Image).filter(models.Image.id == image_id).first()
-    db.delete(db_image)
-    db.commit()
+    if db_image:
+        db.delete(db_image)
+        db.commit()
     return db_image
